Drf_authentication/requirements.txt
asgiref==3.8.1
django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2